requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
easyocr
sentence-transformers
